+++ server/plugins/playwright-cleanup.ts
@@ -1,0 +1,7 @@
+import { closeLibellePlaywrightBrowser } from '../services/recipe-ingestion/fetchLibelleRecipePagePlaywright'
+
+export default defineNitroPlugin((nitroApp) => {
+  nitroApp.hooks.hook('close', async () => {
+    await closeLibellePlaywrightBrowser()
+  })
+})